lara/col/climb: separate water routines
This splits water related collision routines into their own module.
